Break type result out of util
